requirements.txt
openai>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
pyyaml>=6.0
scipy>=1.7.0
pypinyin>=0.46.0
python-Levenshtein>=0.12.2
jieba>=0.42.1
tqdm>=4.62.0
xlrd>=2.0.1
openpyxl>=3.0.9
